rename routines, delete useless ones
